SystemServer/trainning-server/requirements.txt
-i https://pypi.org/simple
aiofiles==0.4.0
aiohttp==4.0.0a0
async-timeout==3.0.1
attrs==19.1.0
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
click==7.0
face-recognition-models==0.3.0
face-recognition==1.2.3
google-api-core==1.8.2
google-auth==1.6.3
google-cloud-core==0.29.1
google-cloud-storage==1.14.0
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0b9
idna==2.8
multidict==4.5.2
numpy==1.16.2
opencv-python==4.0.0.21
pillow==5.4.1
protobuf==3.7.1
pyasn1-modules==0.2.4
pyasn1==0.4.5
python-dotenv==0.10.1
python-engineio==3.5.0
python-socketio==4.0.0
pytz==2018.9
requests==2.21.0
rsa==4.0
scikit-learn==0.19.2
scipy==1.2.1
six==1.12.0
urllib3==1.24.1
yarl==1.3.0
